refactor: extract non-worker imgproxy tools into lib/imgproxy.js
